% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlW2L.R
\name{xlW2L}
\alias{xlW2L}
\title{Convert Execl Wide Table to Long Data Frame}
\usage{
xlW2L(filePath, shtNum = 1, celIStart, cellEnd, x1cell, x1Name = "",
  yName = "")
}
\arguments{
\item{filePath}{the Excel file with extension .xlsx, .xls, xlsm.}

\item{shtNum}{an Excel sheet number for data table, defaults to 1.}

\item{celIStart}{an Excel data starting Cell Index (upper left corner), e.g., em218, eM218, EM218, or 218Em, 218em (with max.2-letter column header).}

\item{cellEnd}{an Excel data ending Cell Index (bottom right corner), e.g., dW543 (with max.2-letter column header).}

\item{x1cell}{an Excel Cell Index for cell with key header, which separates identification variables from Y-value varables, e.g., bf38 (with max.2-letter column header).}

\item{x1Name}{optional key header, defaults to "" will use the original Excel key name if it exists.}

\item{yName}{optional response variable name for y-values returned in Long Data Frame, defaults to "" will use "Y" for header of y-value column in Long Data Frame.}
}
\value{
xlW2L(..)$x1T or xlW2L(..)[1] -- the original Excel table;  xlW2L(..)wideF or xlW2L(..)[2] -- the wide form after cleaning up;  xlW2L(..)longF or xlW2L(..)[3] -- the long data form.
}
\description{
This function converts Execl Wide Table to Long Data Frame.
}
\examples{
xlW2L(filePath,,"e8","o44","i8")$longF
}
\keyword{Excel}
